[#17] added user profile screen ui (#21)
* [#17] added profile screen Ui component

* [#17] added user profile screen

* [#17] minor ui enhancement on user profile screen

* [#17] routine push

* [#17] fixes master merge conflict

* [#17] added profile actions button added

* [#17] added user profile screen UI

* [#17] removed unnecessary code

* [#17] Fix: Updated actions/upload-artifact to v4

+++ app/src/main/java/np/com/naxa/drone_tasking_manager/features/user/profile/dto/UserProfileDto.kt
@@ -2,6 +2,7 @@ package np.com.naxa.drone_tasking_manager.features.user.profile.dto
 
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigInteger
 
 data class UserProfileDto(
     @SerializedName("certificate_file") val certificateFile: String?,
@@ -26,5 +27,5 @@ data class UserProfileDto(
     @SerializedName("registration_certificate_url") val registrationCertificateUrl: String?,
     @SerializedName("registration_file") val registrationFile: String?,
     @SerializedName("role") val role: List<String?>?,
-    @SerializedName("user_id") val userId: Long?
+    @SerializedName("user_id") val userId: BigInteger?
 )
